Revert "Rename internal to network"
This reverts commit 9a846a426ed3c4e38f841b7e004466309cb6e45d.
